Make this implementation keep compliance with spec
EGL spec 1.5, 3.12: Resources explicitly allocated by calls to EGL, such as contexts, surfaces, and
configuration lists, are not affected by eglReleaseThread. Such resources belong
not to the thread, but to the EGL implementation as a whole.
EGL is now c++ project with C API
